Highlights inpaint iop: add midtones threshold
The clipped highlights are likely to have colors
like the other [non-clipped] highlights, not
low-light parts of image.

Thus, we need to separate the values that we will
use for inpainting not only from the top (clip),
but from the bottom.

Any values that are within this (midtones, clip)
range are going to be used for inpainting.