IOP API: add process_prepare{,_cl}().
The idea is to simplify the life of global iops like
colorreconstruction, ashift, etc.

They _need_ the whole input buffer, not just some part of it
they'd get e.g. in full pipe during darkroom, or in tiling context.

Also, there is some interest in sampling something other than
the current ivoid, so the position where this new callback would
be called will hopefully be configurable...